requirements.txt
kafka-python>=2.0.5,<2.3
